'vets1.0'

+++ app/Services/Transfer/TransferService.php
@@ -13,7 +13,8 @@ use App\Models\Order;
 use App\Models\PreOrderCar;
 use App\Models\TransferDeal;
 use App\Models\TransferOrder;
-use App\Services\AuthenticationService;
+use App\Services\AuthService;
+use App\Services\Client\ClientService;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class TransferService
@@ -21,7 +22,7 @@ class TransferService
 
     public function getCollection($request)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         if ($user && $user->role === 'liner') {
             $orders = TransferOrder::where('closed', 0);
             return TransferOrderResource::collection($orders->orderByDesc('date')->get());
@@ -30,7 +31,7 @@ class TransferService
 
     public function getCurrenCollection($request)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         if ($user && $user->role === 'liner') {
             $orders = TransferOrder::where('owner_liner_id', $user->id)->orWhere('recipient_liner_id', $user->id);
             return TransferOrderResource::collection($orders->orderByDesc('date')->get());
@@ -39,7 +40,7 @@ class TransferService
 
     public function getById($id)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         if ($user && $user->role === 'liner') {
             $order = TransferOrder::find($id);
 
@@ -55,14 +56,14 @@ class TransferService
 
     public function pfs($request)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         $order = TransferOrder::find($request->transfer_order_id);
         return $order;
     }
 
     public function store($request)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         $order = Order::find($request->order_id);
         $transfer = TransferOrder::where('order_id', $order->id)->first();
         if (!$transfer) {
@@ -90,7 +91,7 @@ class TransferService
 
     public function getCollectionDeal($request)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         if ($user && $user->role === 'liner') {
             $transfer_order = TransferOrder::find($request->transfer_order_id);
             if ($transfer_order->owner_liner_id === $user->id) {
@@ -102,13 +103,21 @@ class TransferService
 
     public function storeDeal($request)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         $transfer_order_id = $request->transfer_order_id;
         $amount = $request->amount;
 
         $transferOrder = TransferOrder::find($transfer_order_id);
 
-        if ($amount) {
+        $client = Client::where('idnum', $request->client['idnum'])->first();
+
+        if ($client) {
+            $client = app(ClientService::class)->update($request->client, $client->id);
+        } else {
+            $client = app(ClientService::class)->create($request->client);
+        }
+
+        if ($amount && $client) {
             if ($user->id !== $transferOrder->ownner_liner_id) {
                 $deal = new TransferDeal;
                 $deal->liner_id = $user->id;
@@ -119,14 +128,14 @@ class TransferService
 
                 return ['status' => true];
             }
-        }else{
+        } else {
             return ['status' => false];
         }
     }
 
     public function acceptDeal($id): array
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         $transferDeal = TransferDeal::find($id);
         $transferOrder = $transferDeal->transfer_order;
 
@@ -136,18 +145,18 @@ class TransferService
             $transferOrder->closed = 1;
             $transferOrder->save();
             return ['status' => true];
-        }else{
+        } else {
             return ['status' => false];
         }
     }
 
     public function closeDeal($id): array
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         $transferDeal = TransferDeal::find($id);
         $transferOrder = $transferDeal->transfer_order;
 
-        if($user) {
+        if ($user) {
             if ($user->id === $transferOrder->owner_liner_id) {
                 $transferOrder->transfer_deal_id = null;
                 $transferOrder->recipient_liner_id = null;
@@ -162,7 +171,7 @@ class TransferService
 
     public function sign($request): array
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         $sign = $request->sign;
         $transfer_order_id = $request->transfer_order_id;
 
@@ -172,63 +181,67 @@ class TransferService
         $client = Client::where('idnum', $recipient_liner->idnum)->first();
         $order = Order::find($transfer_order->order->id);
 
-        if ($user->role === 'liner') {
+        if ($sign) {
+            if ($user->role === 'liner') {
 
-            if ($transfer_order->owner_liner_id === $user->id) {
-                $transfer_order->owner_sign = $sign;
-                $transfer_order->owner_sign_time = time();
-            }
+                if ($transfer_order->owner_liner_id === $user->id) {
+                    $transfer_order->owner_sign = $sign;
+                    $transfer_order->owner_sign_time = time();
+                }
 
-            if ($transfer_order->recipient_liner_id === $user->id) {
-                $transfer_order->recipient_sign = $sign;
-                $transfer_order->recipient_sign_time = time();
-                $transfer_order->closed = 2;
-                $car = Car::where('order_id', $transfer_order->order_id)->first();
-                $car->cert_title = $client->title;
-                $car->cert_idnum = $client->idnum;
-                $car->save();
+                if ($transfer_order->recipient_liner_id === $user->id) {
+                    $transfer_order->recipient_sign = $sign;
+                    $transfer_order->recipient_sign_time = time();
+                    $transfer_order->closed = 2;
 
-                $order->blocked = 0;
-                $order->client_id = $client->id;
-                $order->save();
-            }
+                    $preorder = PreOrderCar::where('order_id', $transfer_order->order_id)->first();
+                    $preorder->client_id = $client->id;
+                    $preorder->liner_id = $recipient_liner->id;
+                    $preorder->save();
 
-            if (!$client) {
-                $client = new Client;
-                $client->title = json_decode($recipient_liner->profile)->fln;
-                $client->idnum = $recipient_liner->idnum;
-                $client->save();
-            }
+                    $car = Car::where('order_id', $transfer_order->order_id)->first();
+                    $car->cert_title = $client->title;
+                    $car->cert_idnum = $client->idnum;
+                    $car->save();
 
-            if ($transfer_order) {
-
-                $preorder = PreOrderCar::where('order_id', $transfer_order->order_id)->first();
-                $preorder->client_id = $client->id;
-                $preorder->liner_id = $recipient_liner->id;
-                $preorder->save();
-
-                if ($transfer_order->owner_sign != '' && $transfer_order->recipient_sign != '') {
-                    $order = Order::find($transfer_order->order_id);
+                    $order->blocked = 0;
                     $order->client_id = $client->id;
                     $order->save();
                 }
-                $transfer_order->save();
+
+                if (!$client) {
+                    $client = new Client;
+                    $client->title = json_decode($recipient_liner->profile)->fln;
+                    $client->idnum = $recipient_liner->idnum;
+                    $client->save();
+                }
+
+                if ($transfer_order) {
+                    $transfer_order->save();
+                    return ['Сделка подписана'];
+                }
             }
         }
+        return ['Ошибка'];
 
-        return ['Сделка подписана'];
     }
 
     public function close($request)
     {
-        $user = app(AuthenticationService::class)->auth();
+        $user = app(AuthService::class)->auth();
         $transfer_order_id = $request->id;
 
         $transferOrder = TransferOrder::find($transfer_order_id);
+        $order = Order::find($transferOrder->order_id)->first();
 
-        if ($user->id === $transferOrder->owner_liner_id && $transferOrder->closed !== 2) {
-            $transferOrder->delete();
-            return ['закрыта'];
+        if($order){
+            $order->blocked = 0;
+            if($order->save()) {
+                if ($user->id === $transferOrder->owner_liner_id && $transferOrder->closed !== 2) {
+                    $transferOrder->delete();
+                    return ['закрыта'];
+                }
+            }
         }
     }
 }
